requirements.txt
astroid==2.0.4
autopep8==1.3.5
certifi==2018.8.13
colorama==0.3.9
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
Pillow==5.2.0
psycopg2==2.7.5
pycodestyle==2.4.0
pylint==2.1.1
pytz==2018.5
six==1.11.0
wincertstore==0.2
wrapt==1.10.11
